+++ tp.integrador/src/main/java/arq/web/tp/integrador/users/service/impl/UserServiceImpl.java
@@ -8,7 +8,9 @@ import arq.web.tp.integrador.users.entity.UserEntity;
 import arq.web.tp.integrador.users.service.UserService;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.util.ObjectUtils;
 
 
 import java.util.List;
@@ -32,7 +34,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserDTO getUser(Long userId) throws CustomException {
-        UserEntity entity = userJPARepository.getReferenceById(userId);
+        UserEntity entity = userJPARepository.getUserById(userId);
+        if (entity == null || ObjectUtils.isEmpty(entity)) {
+            throw new CustomException("User not found", HttpStatus.NOT_FOUND);
+        }
         return UserConverter.toDTO(entity);
     }
 
@@ -40,12 +45,20 @@ public class UserServiceImpl implements UserService {
     @Transactional
     public Long createUser(UserDTO userDTO) throws CustomException {
         UserEntity entity = UserConverter.toEntity(userDTO);
-        var r = userJPARepository.save(entity);
-        return r.getId();
+        try {
+            var r = userJPARepository.save(entity);
+            return r.getId();
+        } catch (Exception e) {
+            throw new CustomException("Error creating user", HttpStatus.BAD_REQUEST);
+        }
     }
 
     @Override
     public void deleteUser(Long userId) throws CustomException {
+        var user = userJPARepository.findById(userId);
+        if (user == null || user.isEmpty()) {
+            throw new CustomException("User not found", HttpStatus.NOT_FOUND);
+        }
         userJPARepository.deleteById(userId);
     }
 }
